membuat fitur berdasarkan kategori di tampilan user

+++ routes/web.php
@@ -14,43 +14,44 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-Route::get('/', function () {
-    return view('home');
-});
+// Resep
+Route::get('/', 'UserResepController@index');
+Route::get('/resep-makanan/{slug}', 'UserResepController@showResep');
+Route::get('/resep-makanan-detail/{slug}', 'UserResepController@detailResep');
 
 Auth::routes();
 
-// Route::get('/home', 'HomeController@index')->name('home');
+//  ===============================    ADMIN ======================================
 
 // kategori
-Route::resource('/kategori', 'kategoriController');
+Route::resource('/kategori', 'kategoriController')->middleware('auth');
 
 // resep
-Route::get('/resep', 'ResepController@index');
-Route::get('resep/create', 'ResepController@create');
-Route::post('resep', 'ResepController@store');
-Route::get('resep/{id}/edit', 'ResepController@edit');
-Route::put('resep/{id}', 'ResepController@update');
-Route::delete('resep/{id}', 'ResepController@destroy');
+Route::get('/resep', 'ResepController@index')->middleware('auth');
+Route::get('/resep/create', 'ResepController@create')->middleware('auth');
+Route::post('resep', 'ResepController@store')->middleware('auth');
+Route::get('resep/{id}/edit', 'ResepController@edit')->middleware('auth');
+Route::put('resep/{id}', 'ResepController@update')->middleware('auth');
+Route::delete('resep/{id}', 'ResepController@destroy')->middleware('auth');
 
 // gambar resep
-Route::get('resep/{id}/gambar-makanan','ResepController@tambahGambar');
-Route::post('resep/gambar/{id}','ResepController@storeGambar');
-Route::delete('resep/gambar/{id}', 'ResepController@removeGambar');
+Route::get('resep/{id}/gambar-makanan','ResepController@tambahGambar')->middleware('auth');
+Route::post('resep/gambar/{id}','ResepController@storeGambar')->middleware('auth');
+Route::delete('resep/gambar/{id}', 'ResepController@removeGambar')->middleware('auth');
 
 // bahan makanan
-Route::get('resep/{id}/bahan-makanan', 'ResepController@tambahbahan');
-Route::post('resep/bahan/{id}', 'ResepController@storeBahan');
-Route::get('resep/bahan/{id}/edit', 'ResepController@editBahan');
-Route::put('resep/bahan/{id}', 'ResepController@updateBahan');
-Route::delete('resep/bahan/{id}', 'ResepController@removeBahan');
+Route::get('resep/{id}/bahan-makanan', 'ResepController@tambahbahan')->middleware('auth');
+Route::post('resep/bahan/{id}', 'ResepController@storeBahan')->middleware('auth');
+Route::get('resep/bahan/{id}/edit', 'ResepController@editBahan')->middleware('auth');
+Route::put('resep/bahan/{id}', 'ResepController@updateBahan')->middleware('auth');
+Route::delete('resep/bahan/{id}', 'ResepController@removeBahan')->middleware('auth');
 
 // cara membuat makanan
-Route::get('resep/{id}/cara-membuat', 'ResepController@tambahCaraMembuat');
-Route::post('resep/cara-membuat/{id}', 'ResepController@storeCaraMembuat');
-Route::get('resep/cara-membuat/{id}/edit', 'ResepController@editCaraMembuat');
-Route::put('resep/cara-membuat/{id}', 'ResepController@updateCaraMembuat');
-Route::delete('resep/cara-membuat/{id}', 'ResepController@removeCaraMembuat');
+Route::get('resep/{id}/cara-membuat', 'ResepController@tambahCaraMembuat')->middleware('auth');
+Route::post('resep/cara-membuat/{id}', 'ResepController@storeCaraMembuat')->middleware('auth');
+Route::get('resep/cara-membuat/{id}/edit', 'ResepController@editCaraMembuat')->middleware('auth');
+Route::put('resep/cara-membuat/{id}', 'ResepController@updateCaraMembuat')->middleware('auth');
+Route::delete('resep/cara-membuat/{id}', 'ResepController@removeCaraMembuat')->middleware('auth');
 
 // artikel
-Route::resource('/artikel', 'ArtikelController');
+Route::resource('/artikel', 'ArtikelController')->middleware('auth');
